pyproject.toml
[project]
name = "predictive_dialer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Taha Nadeem Khan", email = "tahak6884@gmail.com" },
]
requires-python = ">=3.12"
dependencies = [
    "openai>=1.12.0",
    "openai-agents>=0.0.14",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sqlmodel>=0.0.24",
]

[project.scripts]
add_data = "predictive_dialer.db:add_sample_data"
display_data = "predictive_dialer.db:display_calls"
agents = "predictive_dialer.agents1:main"
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
